Allow html-styling for Android strings by using double quotes convention
Android has two ways of adding html-styling:
`<![CDATA[styled text]]>`
`<a href=“/item”>styled text</a>`
so now twine will surround text with ``text`` to show that it is to be left as is and outputted verbatim to whatever’s inside the double quotes. iOS strings do not support such extensive HTML-styling in their string resources so we do not include double quoted strings to be included in iOS string files.