Exclude definitions that are not translated in the language
Twine automatically inserts the base language’s definition into the foreign language, but we do not want this behaviour in Android because it will automatically pull from the base language’s value in code. This will also show up as red/missing translation in Android Studio’s translation editor, which will notify the developer that they are missing translations, instead of silently filling it in.
This shouldn’t happen on iOS because without that replacement definition, the system will not automatically use the base language’s value and show the key instead, which is in snake case and is ugly.